Refactor: [FE] build error 해결
